Updating Parsons Fork!

+++ parsons/databases/redshift/rs_copy_table.py
@@ -54,7 +54,7 @@ class RedshiftCopyTable(object):
         if blanksasnull:
             sql += "blanksasnull \n"
         if nullas:
-            sql += f"nullas {nullas}"
+            sql += f"null as {nullas}"
         if acceptinvchars:
             sql += "acceptinvchars \n"
         if truncatecolumns:
@@ -117,4 +117,6 @@ class RedshiftCopyTable(object):
         return key
 
     def temp_s3_delete(self, key):
-        self.s3.remove_file(self.s3_temp_bucket, key)
+
+        if key:
+            self.s3.remove_file(self.s3_temp_bucket, key)
